refactor(character, combat): replace Resource with Gauge and refactor related logic
- Replaced `Resource` class with a new `Gauge` record for better immutability.
- Updated `Health` to use `Gauge`, including validation, add, and subtract methods.
- Refactored `Damage` to use integers instead of doubles for consistency.
- Applied changes to combat logic, mappers, DTOs, and tests.
- Adjusted database schema to reflect health value as integers instead of doubles.
- Updated monsters.json to align with the new health and gauge structure.

+++ app/src/main/java/com/arcathoria/monster/dto/MonsterDTO.java
@@ -2,5 +2,5 @@ package com.arcathoria.monster.dto;
 
 import java.util.UUID;
 
-public record MonsterDTO(UUID id, String name, Double currentHealth, Double maxHealth, Integer intelligence) {
+public record MonsterDTO(UUID id, String name, Integer currentHealth, Integer maxHealth, Integer intelligence) {
 }
