+++ domain/src/main/java/com/arcathoria/combat/MeleeMagicDamageStrategy.java
@@ -3,7 +3,7 @@ package com.arcathoria.combat;
 class MeleeMagicDamageStrategy implements DamageCalculator {
 
     @Override
-    public double calculate(final Participant participant) {
+    public int calculate(final Participant participant) {
         return (8 + (participant.getIntelligenceLevel() - 1) * 2);
     }
 }
