+++ infrastructure/src/main/resources/db/changelog/2025-04-14/03-add-health-column-in-characters-table.sql
@@ -2,5 +2,5 @@
 
 --changeset krzysztof:3
 ALTER TABLE characters
-ADD COLUMN max_health DOUBLE PRECISION;
+ADD COLUMN max_health INT;
